Moved Components Back to Component Dir
